@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "com.costs.newcosts"
         minSdkVersion 14
-        targetSdkVersion 26
-        versionCode 26
+        targetSdkVersion 29
+        versionCode 27
         versionName "1.151"
     }
     buildTypes {
@@ -31,6 +31,12 @@ android {
         exclude 'META-INF/notice.txt'
         exclude 'META-INF/ASL2.0'
     }
+    configurations {
+        all {
+            exclude module: 'httpclient'
+            exclude module: 'commons-logging'
+        }
+    }
 }
 
 dependencies {
@@ -49,8 +55,8 @@ dependencies {
 
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support:design:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
 
 //    implementation 'com.google.android.gms:play-services-auth:10.0.1'
 //    implementation 'com.google.android.gms:play-services-drive:10.0.1'
