@@ -1,8 +1,8 @@
 package com.costs.newcosts;
 
 import android.database.Cursor;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
