@@ -2,7 +2,7 @@ package com.costs.newcosts.common.sms.provider;
 
 import android.content.Context;
 import android.database.Cursor;
-import android.support.v4.content.CursorLoader;
+import androidx.loader.content.CursorLoader;
 
 import com.costs.newcosts.common.sms.reader_result.SmsExpenseReaderResult;
 
